+++ bikespace_dashboard/js/app.js
@@ -1,11 +1,11 @@
 import {PanelNav} from './components/sidebar/panel_nav.js';
 import {SharedState, Component} from './components/main.js';
 import {Map} from './components/map.js';
-import {SummaryBox} from './components/sidebar/summary_box.js';
-import {IssueChart} from './components/sidebar/issue_chart.js';
-import {DayChart} from './components/sidebar/day_chart.js';
-import {DurationTimeOfDayChart} from './components/sidebar/duration_tod_chart.js';
-import {Submissions} from './components/sidebar/submissions.js';
+import {SummaryBox} from './components/sidebar/data-panel/summary_box.js';
+import {IssueChart} from './components/sidebar/data-panel/issue_chart.js';
+import {DayChart} from './components/sidebar/data-panel/day_chart.js';
+import {DurationTimeOfDayChart} from './components/sidebar/data-panel/duration_tod_chart.js';
+import {Submissions} from './components/sidebar/feed/submissions.js';
 
 $.ajax({
   url: 'https://api-dev.bikespace.ca/api/v2/submissions?limit=5000',
