+++ bikespace_dashboard/js/components/sidebar/data-panel/duration_tod_chart.js
@@ -1,4 +1,4 @@
-import {Component} from '../main.js';
+import {Component} from '../../main.js';
 import {defaults, cssVarHSL} from './plot_utils.js';
 
 // Note: "tod" = "Time of Day"
@@ -9,7 +9,7 @@ class DurationTimeOfDayChart extends Component {
    * @param {string} parent JQuery selector for parent element
    * @param {string} root_id tag id for root div
    * @param {Object} shared_state
-   * @param {import('../main.js').ComponentOptions} [options = {}] Options for the component
+   * @param {import('../../main.js').ComponentOptions} [options = {}] Options for the component
    */
   constructor(parent, root_id, shared_state, options = {}) {
     super(parent, root_id, shared_state, options);
