+++ bikespace_dashboard/js/components/sidebar/data-panel/plot_utils.js
