+++ bikespace_dashboard/js/components/sidebar/data-panel/summary_box.js
@@ -1,4 +1,4 @@
-import {Component} from '../main.js';
+import {Component} from '../../main.js';
 
 class SummaryBox extends Component {
   /**
@@ -6,7 +6,7 @@ class SummaryBox extends Component {
    * @param {string} parent JQuery selector for parent element
    * @param {string} root_id tag id for root div
    * @param {Object} shared_state
-   * @param {import('../main.js').ComponentOptions} [options = {}] Options for the component
+   * @param {import('../../main.js').ComponentOptions} [options = {}] Options for the component
    */
   constructor(parent, root_id, shared_state, options = {}) {
     super(parent, root_id, shared_state, options);
